Phase 2 of #827: extract Alpine factory from rule_detail (#843)
Phase 2 of #827 — extract `rule_detail`'s Alpine factory to follow the page-component convention codified in #828.

This is **the first Source-C migration** (props-templated Go function returning `<script>` HTML), so it puts the `@templ.JSONScript` + `data-*` attribute split through its first real workout. Every dynamic value the old `ruleDetailBootstrap(p)` interpolated now flows through one of the two convention-blessed channels — there is no `fmt.Sprintf` in Go for JS source any more, and `static/js/admin/components/rule_detail.js` is 100% static.

## How this PR exercises the JSONScript + data-attr pattern

The old `ruleDetailBootstrap(p)` had three interpolation sites. Each is mapped to the right channel:

| Source-today | Shape | New channel |
|---|---|---|
| `window.__bbCategories = {{toJSON .Categories}}` | Typed slice (`[]service.CategoryResponse`) | `@templ.JSONScript("rule-detail-data", p.Categories)` parsed once in `init()` |
| `/-/rules/{{.Rule.ID}}/apply` | Scalar URL | `data-apply-url` on the `x-data` root, read via `this.$el.dataset.applyUrl` |
| `/-/rules/{{.Rule.ID}}/toggle` | Scalar URL | `data-toggle-url` on the `x-data` root, read via `this.$el.dataset.toggleUrl` |

The factory now takes no arguments; `x-data="ruleDetail"` is the string-literal form the lint requires.

### Wrinkle worth noting (covered, not new convention)

`tx_row.templ` renders an inline `categoryPicker({categories: window.__bbCategories, ...})` for every Recent Application and Matching Transaction row on this page (~18 invocations). The shared component reaches into `window.__bbCategories` directly — no Alpine scope, no JSONScript awareness. To keep that contract intact, `init()` parses the JSON script **and** assigns the parsed value to `window.__bbCategories`. Same data, two consumers; the JSON is parsed once.

`showToast` and `quickSetCategory` (consumed by `tx_row`'s inline `x-init` watcher) remain at module-top-level on `window` for the same reason. The minimal no-op `bulk` store stays inside the `alpine:init` listener.

This is a back-compat shim, not a new pattern. The convention spec in `docs/design-system.md` already covers "share JSONScript-parsed data with a non-Alpine consumer via window" implicitly via the "complex initial state → JSONScript" guidance — no doc changes needed.

## What changed

- **New file:** `static/js/admin/components/rule_detail.js` — wraps the factory in `Alpine.data('ruleDetail', () => ({ init() { ... }, ... }))`, parses `#rule-detail-data` JSON, reads URLs from `this.$el.dataset`. Synchronously loaded `<script src>` at the top of the templ component (no `defer`).
- **Deleted:** `rule_detail_scripts.go` (167 lines, gone). The orphaned `ruleDetailCategoriesJSON` helper goes with it.
- **`rule_detail.templ`:** drops `@templ.Raw(ruleDetailBootstrap(p))`, adds the `<script src>`, adds `@templ.JSONScript("rule-detail-data", p.Categories)`, switches `x-data="ruleDetail()"` → `x-data="ruleDetail"`, adds `data-apply-url` / `data-toggle-url`.
- **`rule_detail_types.go`:** drops `ruleDetailCategoriesJSON` and the `encoding/json` import; adds a tiny `ruleID(*service.TransactionRuleResponse) string` helper used by the templ to build the URL data-attrs safely when the rule pointer is nil.

## Lint ceiling

Re-measured the largest inline `<script>` block across `*.templ` after this change: `connections.templ:454-620` is still the max at **147 content lines**. No change — the existing `inlineScriptCeiling = 180` covers it. (My change deletes a Go-string `<script>` factory, which the lint already excluded; nothing crosses the watermark.)

## Browser validation

Validated at `/rules/<id>` (the "Auto-tag new transactions for review" system rule, 8 pending matches):

- Rule details render: header, what-it-does, stat tiles, recent applications, pending matches, sync log
- Inline `categoryPicker` on tx-rows is wired (18 picker `x-data` instances reading from `window.__bbCategories`, which has 17 categories)
- Apply modal opens via `openApplyModal()`, closes via Cancel — Alpine state mutations work
- Reactive bindings on `applyUrl` / `toggleUrl` from `data-*` confirmed via DevTools eval:
  ```
  applyUrl:  /-/rules/<id>/apply
  toggleUrl: /-/rules/<id>/toggle
  ```
- `bulk` store registered, `quickSetCategory` and `showToast` available on `window`
- Console silent (only standard browser warnings unrelated to this change)

<img src="https://i.img402.dev/anwjiffv0r.jpg" width="800" alt="rule_detail — after extraction">

## Test plan

- [x] `templ generate`, `templ fmt .` clean
- [x] `go build ./...` clean
- [x] `go vet ./...` clean
- [x] `go test ./...` passes (including `TestNoLargeInlineScripts`)
- [x] Browser-validated at `/rules/<id>` — page renders, modal works, no console errors
- [x] `window.__bbCategories` populated for tx-row category pickers (back-compat shim verified)
